Commiting unit test passed version of code, changes to deployment strategy for the spss*.dll's in general.
Inclusion of new license and documentation in deployment.

Change of directories from conventional x86/x64 naming to win32/win64 naming
this is to be inline with the directory struture in the IBM released SPSSIO zip
files, it should mean that any future version is easier to incorporate just by
unzipping the ibm file into the resources directory of the solution.